+++ CHANGELOG.md
@@ -59,6 +59,8 @@
 
 ### v3.1.16 / 2020-04-2x
 
+* [324](https://github.com/Vanessa219/vditor/issues/324) 支持多款主题预览 `引入特性`
+
 ### v3.1.15 / 2020-04-21
 
 * [333](https://github.com/Vanessa219/vditor/issues/333) 重新设计帮助菜单 `改进功能`
@@ -112,8 +114,8 @@
   * 添加 lazyLoadImageRender 静态方法
   * `setTheme` 方法添加 `codeTheme` 参数
   * insert line 默认快捷键由 `⌘-⇧-D` 修改为 `⌘-⇧-H`，添加下移 `⌘-⇧-D`、上移 `⌘-⇧-U` 快捷键，移除上传、编辑器模式切换快捷键
-  * 为 `options.toolbar` 添加 outdent，indent, outline, insert-after, insert-before, 子菜单，code-theme
   * 为 `options.toolbar` 添加 toolbar 参数，最多可进行 3 级菜单
+  * 为 `options.toolbar` 添加 outdent，indent, outline, insert-after, insert-before, 子菜单，code-theme, content-theme
 
 ### v3.0.12 / 2020-04-06